requirements.txt
requests==2.31.0
envfileparser==0.0.8
luma.oled==3.13.0